Updated Query and Removed Reporting on State case duplicates
Our client gave us an updated query to use with the state database. Thus, I have made the necessary changes to the compare.py script which checks the add_time field to make sure we are comparing the most recent version of the case stored in the state database with the cdc file.